Created a TablePage component and Rendered it in App.jsx
